+++ .gitignore
@@ -17,3 +17,4 @@ web-build/
 .metro-health-check*
 .env
 eas.json
+google-services.json
