+++ app/(main)/alert.jsx
@@ -1,15 +1,12 @@
 import { useState } from "react";
 import { View } from "react-native";
 import { ActivityIndicator, Button, Text, IconButton, ProgressBar } from "react-native-paper";
-import { Audio } from "expo-av";
-import { decode } from 'base64-arraybuffer';
-import * as Notifications from 'expo-notifications';
-import * as FileSystem from "expo-file-system";
 
 import { viewStyle } from "../../ui/style";
 import { useAuth } from "../../contexts/auth";
 import { supabase } from "../../lib/supabase";
 import { useLocation } from "../../contexts/location";
+import { useRecorder } from "../../contexts/recording";
 
 /**
  * Alert Page for the app. This will be the main page for the app, and is the first
@@ -19,56 +16,9 @@ import { useLocation } from "../../contexts/location";
 export default function AlertPage() {
     const { loggedIn } = useAuth();
     const { location } = useLocation();
-    const [ recording, setRecording ] = useState(null);
+    const { startRecording, stopRecording, recording } = useRecorder();
     const [ statusMessage, setMessage ] = useState("Awaiting input...");
     const [ loading, setLoading ] = useState(false);
-
-    async function startRecording() {
-        try {
-            console.log('Requesting permissions..');
-            await Audio.requestPermissionsAsync();
-            await Audio.setAudioModeAsync({
-                allowsRecordingIOS: true,
-                playsInSilentModeIOS: true,
-            });
-    
-            console.log('Starting recording..');
-            await Notifications.scheduleNotificationAsync({
-                content: {
-                    title: "Recording in progress...",
-                    subtitle: "Tap to stop the recording.",
-                    data: { onPress: stopRecording },
-                    autoDismiss: true,
-                },
-                identifier: "recording",
-                trigger: null
-            });
-            const { recording } = await Audio.Recording.createAsync( Audio.RecordingOptionsPresets.HIGH_QUALITY);
-            setRecording(recording);
-            console.log('Recording started');
-        } catch (err) {
-            console.error('Failed to start recording', err);
-        }
-    }
-
-    async function stopRecording() {
-        console.log('Stopping recording..');
-        setRecording(undefined);
-        await recording.stopAndUnloadAsync();
-        await Audio.setAudioModeAsync({
-          allowsRecordingIOS: false,
-        });
-        const uri = recording.getURI();
-        console.log('Recording stopped and stored at', uri);
-        const file = await FileSystem.readAsStringAsync(uri, { encoding: FileSystem.EncodingType.Base64 });
-        const { data, error } = await supabase.storage.from("audio").upload(`${loggedIn.id}/${Date.now()}.m4a`, decode(file), {
-            contentType: 'audio/mpeg'
-        });
-        if (error) {
-            console.log("failed to upload. " + error.message);
-        }
-        console.log(data);
-    }
 
     const handleDistress = async () => {
         setLoading(true);
